requirements.txt
fastapi
uvicorn
google-generativeai
google-genai
Pillow
requests
python-dotenv
jinja2
